feat(toolbar): add copy to clipboard button
This feature allows you to copy the colors of the canvas to your clipboard, so you can take these colors into your video game of choice and do cool things with them!

In particular, the colors that we return are actually just pointers to the index of the color as it appears in the palette.